Update patch set 3
Patch Set 3:

The metadata of the parsed VNFD should be taken from the metadata of the CSAR. The current implementation in toscaparser/vnfdmodel uses the properties of the first node_template to populate the csarVersion, vendor, version fields. There is no guarantee in Tosca that the first node_template has the csarVersion, vendor, version attributes. These metadata should be taken from deep copying the metadata section of the CSAR. The current implementation set the csarVersion, vendor, version fields to undefined even if they were specified as metadata if the first node does not have that properties.
This also needs prior fixes in where the VNFD is converted into structures that is understood by SDC. That conversion also looses the metadata during converting the VF-C based VNFD into SDC format.

Patch-set: 3
Label: Code-Review=0