Update patch set 1
Patch Set 1:

> The GVNFM use tosca-nfv-v1.0-wd04-rev04 profile. It is not a bug.
 > For the ONAP VNF TOSCA profile is not defined yet. After model
 > subcommittee defined, we will consider how to support it.

This code is not related to the GVNFM. The parser in the catalog parses the Tosca template. When ONAP send the template to the catalog it is in the format of the attached file and not  tosca-nfv-v1.0-wd04-rev04. 

 > The GVNFM use tosca-nfv-v1.0-wd04-rev04 profile. It is not a bug.
 > For the ONAP VNF TOSCA profile is not defined yet. After model
 > subcommittee defined, we will consider how to support it.

Patch-set: 1
Reviewer: Gerrit User 2255 <2255@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0